scraper/requirements.txt
# Production Web Scraper Requirements
# Python 3.10+

# Core dependencies
beautifulsoup4>=4.12.0
lxml>=4.9.0
httpx>=0.25.0
playwright>=1.40.0

# Data processing
pandas>=2.0.0

# Logging and CLI
rich>=13.0.0

# Configuration
pyyaml>=6.0

# Utilities
python-dateutil>=2.8.0

